+++ stroom-pipeline/src/main/java/stroom/pipeline/refdata/store/RefDataValue.java
@@ -30,7 +30,7 @@ public abstract class RefDataValue {
      * @return The hashcode of just the underlying value that this object wraps
      * rather than hashcode of the whole object.
      */
-    public abstract int getValueHashCode();
+    public abstract long getValueHashCode(final ValueStoreHashAlgorithm valueStoreHashAlgorithm);
 
     /**
      * @return A code to represent the class of the implementation,