+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Change the off-heap ref store to use xxHash for hashing its values.
+
 
 ## [v7.0-beta.66] - 2020-09-24